+++ test/IntegrationTests/Helpers/TestHelper.cs
@@ -93,7 +93,7 @@ public abstract class TestHelper
         if (testSettings.MetricsSettings != null)
         {
             builder = builder.WithEnvironment("OTEL_EXPORTER_OTLP_ENDPOINT", agentBaseUrl);
-            builder = builder.WithEnvironment("OTEL_AUTO_METRIC_EXPORT_INTERVAL", "1000");
+            builder = builder.WithEnvironment("OTEL_METRIC_EXPORT_INTERVAL", "1000");
             builder = builder.WithEnvironment("OTEL_DOTNET_AUTO_METRICS_ENABLED_INSTRUMENTATIONS", "AspNet");
         }