+++ CHANGELOG.md
@@ -29,6 +29,7 @@ This release is built on top of [OpenTelemetry .NET](https://github.com/open-tel
 - Support for .NET Runtime metrics collection using
   the `OpenTelemetry.Instrumentation.Runtime` package.
 - Support for ASP.NET and HttpClient metrics instrumentations.
+- Support for Prometheus Exporter HttpListener version.
 
 ### Changed